feat: add database versioning
