+++ packages/app/load.envs.js
@@ -10,7 +10,6 @@ function getVersion() {
   );
   return {
     version: packageJson.version,
-    database: packageJson.database,
   };
 }
 
@@ -38,11 +37,9 @@ function getPublicEnvs() {
   );
 }
 
-// Export the version to be used on database
-// and application level
+// Export the version to be used on application level
 const versions = getVersion();
 process.env.PORT = 3000;
 process.env.VITE_APP_VERSION = process.env.VITE_APP_VERSION || versions.version;
-process.env.VITE_DATABASE_VERSION = versions.database;
 
 module.exports.getPublicEnvs = getPublicEnvs;
